linuch werror tips

+++ operations.c
@@ -58,7 +58,7 @@ double* get_operands(handle* const top, unsigned int num_of_operands) {
 	asrt(top->head);
 	double* operands = malloc(num_of_operands * sizeof(double));
 	asrt(operands);
-	for (int i = 0; i < num_of_operands; ++i) {
+	for (size_t i = 0; i < num_of_operands; ++i) {
 		asrt(pop(top, operands + i));
 		if (!(top->head) && (i != num_of_operands - 1)) {
 			free(operands);
